feat!: version API output in cache by generation
This change invalidates all existing cache entries belonging to the extension.

The generation is an integer that should be bumped on changes to the API output structure. As this is being included in v0.8.0, the initial ID is `8`. This is purely for contingency, so maps do not have to be edited before their cache entries expire naturally.