refactor!: handle `filter` in API always at request time
Raising cache generation to `10`.

Instead of producing a blob for each missed `filter` value, we'll now process the markers once for the entire map. Filtering is a very cheap operation and will be now done as a post-processing step (running on every request).

Main blob will be retrieved regardless of the `filter` parameter. Post-processing results are completely uncached.

This eliminates the (unlikely) possibility for system resources to be exhausted through this endpoint, and lowers the number of cache misses.