fix: rename `location` checklist sort to `groupDeclaration`
I accidentally named this wrong in v0.15. New schema version of v0.16, as this is not just a cosmetic change (unfortunately).

There's a hack to permit `location` - schema version is not checked. I've got no plans to remove support for v0.15 yet (though it would help test Screebot), but the hack allows the disallowed value in v0.16. No way to fix this without a significant alteration to the DataModel class or #35.