fix: no need to offset coordinates for tile lookup
Offset is in world-space and should apply to the whole GridLayer instead. No clue how to really do that (the layer does not seem to support custom positioning), but the option should be safe to ditch (have to check wiki.gg first) - we can either switch to old behaviour if it's set, for a transitional period, or just remove the option.

Issue: #220